Contents of `ui/webui` moved to `webui/ui_bundled`
This is just a moving of files about that required correction for
inclusion paths and overrides.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/d236ab1a0320384b0267eec9b6765c666944984d

commit d236ab1a0320384b0267eec9b6765c666944984d
Author: Mike Dougherty <michaeldo@chromium.org>
Date:   Fri Feb 9 05:37:30 2024 +0000

    [Code Health][iOS] Move ios/c/b/ui/webui to ui_bundled

    To simplify the folder hierarchy on iOS and have a clearer layer
    separation, all the files in ios/c/b/ui/<foo>/ will be moved to
    ios/c/b/<foo>/ui_bundled/ to showcase that they are model files.

    Bug: 40223315, 324232870
