+++ ios/browser/ui/webui/BUILD.gn
@@ -24,8 +24,8 @@ source_set("webui") {
     "//components/resources",
     "//ios/chrome/browser/shared/model/browser_state",
     "//ios/chrome/browser/shared/model/url:constants",
-    "//ios/chrome/browser/ui/webui",
-    "//ios/chrome/browser/ui/webui:webui_internal",
+    "//ios/chrome/browser/webui/ui_bundled:coordinator",
+    "//ios/chrome/browser/webui/ui_bundled:webui_internal",
     "//ios/components/webui:url_constants",
     "//ios/web/public/webui",
     "//services/network/public/cpp",
