+++ ios/browser/ui/webui/brave_web_ui_controller_factory.h
@@ -8,7 +8,7 @@
 
 #include <memory>
 
-#import "ios/chrome/browser/ui/webui/chrome_web_ui_ios_controller_factory.h"
+#import "ios/chrome/browser/webui/ui_bundled/chrome_web_ui_ios_controller_factory.h"
 #import "ios/web/public/webui/web_ui_ios.h"
 #import "ios/web/public/webui/web_ui_ios_controller.h"
